Feat: Add Authorization level per user role

+++ backend/src/time-log/time-log.service.ts
@@ -1,14 +1,20 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { CreateTimeLogDto } from './dto/create-time-log.dto';
 import { UpdateTimeLogDto } from './dto/update-time-log.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { GeolocService } from 'src/geoloc/geoloc.service';
+import errors from '../../res/consts';
+import { Role } from 'src/auth/res/roles.enum';
 
 @Injectable()
 export class TimeLogService {
   constructor(private prisma: PrismaService, private geoloc: GeolocService) {}
 
-  async create(createTimeLogDto: CreateTimeLogDto) {
+  async create(createTimeLogDto: CreateTimeLogDto, user: any) {
+    if (user.role == Role.STUDENT && user.id != createTimeLogDto.requestedById) {
+      throw new UnauthorizedException();
+    }
+
     createTimeLogDto.deniedById = null;
     createTimeLogDto.approvedById = null;
 
@@ -50,7 +56,21 @@ export class TimeLogService {
     return response;
   }
 
-  async update(id: string, updateTimeLogDto: UpdateTimeLogDto) {
+  async update(id: string, updateTimeLogDto: UpdateTimeLogDto, user: any) {
+    if (user.role == Role.STUDENT && user.id != id) {
+      throw new UnauthorizedException();
+    }
+
+    const timeLog = await this.prisma.timeLog.findUnique({
+      where: {
+        id: id
+      }
+    });
+
+    if (timeLog.deniedById || timeLog.approvedById) {
+      throw { statusCode: 400, internalCode: 4, message: errors[2] }
+    }
+
     const response = await this.prisma.timeLog.update({
       where: {
         id: id
@@ -60,7 +80,11 @@ export class TimeLogService {
     return { response, message: "Updated" };
   }
 
-  async remove(id: string) {
+  async remove(id: string, user: any) {
+    if (user.role == Role.STUDENT && user.id != id) {
+      throw new UnauthorizedException();
+    }
+
     const response = await this.prisma.timeLog.delete({
       where: {
         id: id
